refactor: resolve all critical/high source code debt
1. Split one-pass.ts (1419→5 lines entry point)
   - cli.ts, config.ts, http.ts for shared infrastructure
   - steps/health.ts, tenant.ts, admin.ts, deploy.ts, participant.ts
   - Same CLI interface, no logic changes

2. Extract route fallback logic into service layer
   - events/[eventId]/route.ts: 177→107 lines, fallback in admin-event-service.ts
   - settings/route.ts: 224→115 lines, fallback in admin-settings-service.ts
   - Route handlers now thin: auth → service → response
