+++ client/Application/app/api/admin/events/[eventId]/route.ts
@@ -8,42 +8,19 @@
  */
 
 import { NextRequest } from 'next/server';
-import { authSkipEnabled } from '@/auth';
 import {
   getAdminSession,
   unauthorizedResponse,
   forbiddenResponse,
   badRequestResponse,
   successResponse,
-  serverApiRequest,
 } from '@/lib/api/server';
-import type { EventDetails, EventStatus } from '@/lib/api/types';
-import { deleteDevEvent, findDevEvent, updateDevEvent } from '../dev-store';
-
-/**
- * イベント更新リクエスト型
- */
-interface UpdateEventRequest {
-  name?: string;
-  slug?: string;
-  description?: string;
-  organizer?: string;
-  eventDate?: string;
-  startTime?: string;
-  endTime?: string;
-  status?: EventStatus;
-  imageUrl?: string;
-}
-
-function isLocalDevFallbackError(error: unknown): boolean {
-  const isAuthSkipUnauthorized =
-    authSkipEnabled &&
-    error instanceof Error &&
-    /^Unauthorized$/i.test(error.message);
-  const isNetworkError =
-    error instanceof TypeError && /fetch failed/i.test(String(error));
-  return isAuthSkipUnauthorized || isNetworkError;
-}
+import {
+  fetchEvent,
+  putEvent,
+  removeEvent,
+} from '@/lib/api/admin-event-service';
+import type { UpdateEventRequest } from '@/lib/api/admin-event-service';
 
 /**
  * GET /api/admin/events/[eventId]
@@ -54,7 +31,6 @@ export async function GET(
   _request: NextRequest,
   { params }: { params: Promise<{ eventId: string }> },
 ) {
-  // 管理者権限チェック
   const session = await getAdminSession();
   if (!session) {
     return session === null
@@ -65,20 +41,9 @@ export async function GET(
   const { eventId } = await params;
 
   try {
-    const data = await serverApiRequest<EventDetails>(
-      `/admin/events/${eventId}`,
-    );
+    const data = await fetchEvent(eventId);
     return successResponse(data);
   } catch (error) {
-    if (isLocalDevFallbackError(error)) {
-      console.warn('Admin event detail fallback to local dev store:', error);
-      const event = findDevEvent(eventId);
-      if (event) {
-        return successResponse(event);
-      }
-    }
-
-    console.error('Failed to fetch event:', error);
     return badRequestResponse(
       error instanceof Error ? error.message : 'Failed to fetch event',
     );
@@ -94,7 +59,6 @@ export async function PUT(
   request: NextRequest,
   { params }: { params: Promise<{ eventId: string }> },
 ) {
-  // 管理者権限チェック
   const session = await getAdminSession();
   if (!session) {
     return session === null
@@ -106,25 +70,9 @@ export async function PUT(
   const body = (await request.json()) as UpdateEventRequest;
 
   try {
-    const data = await serverApiRequest<EventDetails>(
-      `/admin/events/${eventId}`,
-      {
-        method: 'PUT',
-        body: JSON.stringify(body),
-      },
-    );
-
+    const data = await putEvent(eventId, body);
     return successResponse(data);
   } catch (error) {
-    if (isLocalDevFallbackError(error)) {
-      console.warn('Admin event update fallback to local dev store:', error);
-      const event = updateDevEvent(eventId, body);
-      if (event) {
-        return successResponse(event);
-      }
-    }
-
-    console.error('Failed to update event:', error);
     return badRequestResponse(
       error instanceof Error ? error.message : 'Failed to update event',
     );
@@ -147,7 +95,6 @@ export async function DELETE(
   _request: NextRequest,
   { params }: { params: Promise<{ eventId: string }> },
 ) {
-  // 管理者権限チェック
   const session = await getAdminSession();
   if (!session) {
     return session === null
@@ -158,18 +105,9 @@ export async function DELETE(
   const { eventId } = await params;
 
   try {
-    await serverApiRequest<void>(`/admin/events/${eventId}`, {
-      method: 'DELETE',
-    });
-
-    return successResponse({ success: true, message: 'Event deleted' });
+    const data = await removeEvent(eventId);
+    return successResponse(data);
   } catch (error) {
-    if (isLocalDevFallbackError(error) && deleteDevEvent(eventId)) {
-      console.warn('Admin event delete fallback to local dev store:', error);
-      return successResponse({ success: true, message: 'Event deleted' });
-    }
-
-    console.error('Failed to delete event:', error);
     return badRequestResponse(
       error instanceof Error ? error.message : 'Failed to delete event',
     );
